Obsolete non-durable delivery
