+++ src/NServiceBus.Core.Tests/Transports/MulticastTransportOperationTest.cs
@@ -2,6 +2,7 @@ namespace NServiceBus.Core.Tests.Transports
 {
     using System;
     using System.Collections.Generic;
+    using NServiceBus.Performance.TimeToBeReceived;
     using NUnit.Framework;
     using Transport;
 
@@ -14,7 +15,7 @@ namespace NServiceBus.Core.Tests.Transports
             var transportOperation = new MulticastTransportOperation(new OutgoingMessage(Guid.NewGuid().ToString(), new Dictionary<string, string>(), new byte[0]), typeof(Guid));
             var secondTransportOperation = new MulticastTransportOperation(new OutgoingMessage(Guid.NewGuid().ToString(), new Dictionary<string, string>(), new byte[0]), typeof(Guid));
 
-            transportOperation.DeliveryConstraints.Add(new NonDurableDelivery());
+            transportOperation.DeliveryConstraints.Add(new DiscardIfNotReceivedBefore(TimeSpan.FromDays(1)));
 
             Assert.IsEmpty(secondTransportOperation.DeliveryConstraints);
             Assert.IsNotEmpty(transportOperation.DeliveryConstraints);
