+++ src/NServiceBus.Core/DeliveryConstraints/DeliveryConstraint.cs
@@ -1,5 +1,5 @@
 ﻿namespace NServiceBus.DeliveryConstraints
-{   
+{
     /// <summary>
     /// Base class for delivery constraints.
     /// </summary>
